% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepost_estimation_helpers.R
\name{aggregate_saedirect}
\alias{aggregate_saedirect}
\title{Compare estimated and direct parameters in aggregate area levels}
\usage{
aggregate_saedirect(
  ebp_object,
  smp_data,
  pop_data,
  welfare,
  smp_domains,
  pop_domains,
  pop_weights,
  pop_regions,
  smp_regions,
  smp_weights,
  threshold,
  indicator,
  in_sample = TRUE
)
}
\arguments{
\item{ebp_object}{the EBP object produced from by EMDI from unit model estimation
the object is of class "ebp emdi"}

\item{smp_data}{a data.frame; the survey/training dataframe}

\item{pop_data}{data.frame; the population/census dataframe}

\item{welfare}{character; the welfare aggregate variable or outcome variable of interest}

\item{smp_domains}{character; the domain variable within the survey dataframe}

\item{pop_domains}{character; the domain variable within the population/census dataframe}

\item{pop_weights}{character; the population weight variable}

\item{smp_weights}{character; the sample weight variable}

\item{threshold}{numeric; the poverty line or threshold specified}

\item{indicator}{character; an outcome indicator of interest. Options are "Head_Count",
"Gini" and "Poverty_Gap"}

\item{in_sample}{logical, if TRUE only in-sample areas will be used in estimating the 
aggregated EBP model estimates. Otherwise, all areas are included in estimating the means}
}
\description{
The function \code{aggregate_saedirect} computes the head count, poverty gap
and gini estimates at higher levels of aggregation than that estimated by
\code{ebp}. This serves to compare at a more representative level the 
model estimates as well as direct survey estimates for each of the three
parameters.
}
\examples{
data("eusilcA_pop")
data("eusilcA_smp")

#### set of variables used in model estimation
variables <- c("gender", "eqsize", "cash", "self_empl",
              "unempl_ben", "age_ben", "surv_ben",
              "sick_ben", "dis_ben", "rent", "fam_allow",
              "house_allow", "cap_inv", "tax_adj")

### estimate a unit model
emdi_model <- emdiplus::ebp(fixed = as.formula(paste("eqIncome ~ ", paste(variables, 
                                                                         collapse= "+"))),
                           pop_data = eusilcA_pop2, 
                           pop_domains = "district", 
                           smp_data = eusilcA_smp2, 
                           smp_domains = "district",
                           na.rm = TRUE,
                           weights = "weight",
                           pop_weights = "popweights",
                           MSE = TRUE,
                           threshold = 11000,
                           B = 2,
                           L = 2)

 ### estimate aggregated poverty headcounts (EBP and Direct) for regions                            
aggregate_saedirect(ebp_object = emdi_model,
                    smp_data = eusilcA_smp2,
                    pop_data = eusilcA_pop2, 
                    welfare = "eqIncome",
                    smp_domains = "district",
                    pop_domains = "district",
                    pop_weights = "popweights",
                    pop_regions = "state",
                    smp_weights = "weight",
                    threshold = 11000,
                    indicator = "Head_Count",
                    in_sample = FALSE,
                    smp_regions = "state")
                    
 ### estimate aggregated gini (EBP and Direct) for regions                            
aggregate_saedirect(ebp_object = emdi_model,
                    smp_data = eusilcA_smp2,
                    pop_data = eusilcA_pop2, 
                    welfare = "eqIncome",
                    smp_domains = "district",
                    pop_domains = "district",
                    pop_weights = "popweights",
                    pop_regions = "state",
                    smp_weights = "weight",
                    threshold = 11000,
                    indicator = "Gini",
                    in_sample = FALSE,
                    smp_regions = "state")
                    
 ### estimate aggregated Poverty_Gap (EBP and Direct) for regions                            
aggregate_saedirect(ebp_object = emdi_model,
                    smp_data = eusilcA_smp2,
                    pop_data = eusilcA_pop2, 
                    welfare = "eqIncome",
                    smp_domains = "district",
                    pop_domains = "district",
                    pop_weights = "popweights",
                    pop_regions = "state",
                    smp_weights = "weight",
                    threshold = 11000,
                    indicator = "Head_Count",
                    in_sample = FALSE,
                    smp_regions = "state")
                    
 @export
}
