% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepost_estimation_helpers.R
\name{ebp_compute_cv}
\alias{ebp_compute_cv}
\title{Coefficient of Variation (CV) estimations for Unit EBP Model Headcount Estimates}
\usage{
ebp_compute_cv(
  ebp_object,
  smp_data,
  welfare,
  calibvar,
  domainvar,
  boot_type = "calibrate",
  designvar = NULL,
  threshold,
  smp_weights
)
}
\arguments{
\item{ebp_object}{the EBP object produced from by EMDI from unit model estimation
the object is of class "ebp emdi"}

\item{smp_data}{the survey/training data}

\item{welfare}{the welfare aggregate variable or outcome variable of interest}

\item{calibvar}{the calibration variable to be used in method 1}

\item{domainvar}{the target area variable}

\item{boot_type}{the bootstrap type "calibrated" or "naive" to be used in method 1}

\item{designvar}{the survey design variable to be used in estimating the design 
effect for method 3.}

\item{threshold}{the poverty line or threshold specified}

\item{smp_weights}{the weight variable}
}
\description{
Function \code{ebp_compute_cv} estimates CVs for the headcount of the unit model
EBP functions using three different methods. CV, by definition, is the ratio of
mean square error of the head count to the  head count estimates. Therefore, the 
CV types are distinguished by the method of estimating the mean square.
}
\details{
Method 1 uses the calibrated/naive bootstrapping of the MSE which allows to calibrate 
each bootstrap sample on auxiliary information using the \code{direct} function. 
Calibrated bootstrap improves on the bias of the naive bootstrap when used in the 
complex survey context (see \cite{Rao and Wu (1988)}) for more details.

Method 2 employs the Horowitz Thompson variance estimation technique to compute 
MSE i.e. each household is assigned the probability selection within the sample
under a given sampling scheme. The computation employs \code{sae::direct} function.

Method 3 finally uses the design effect adjusted naive calibrated MSE. The design
effect is estimated using the \code{survey::svydesign} function.
}
\examples{
data("eusilcA_pop")
data("eusilcA_smp")

#### set of variables used in model estimation
variables <- c("gender", "eqsize", "cash", "self_empl",
              "unempl_ben", "age_ben", "surv_ben",
              "sick_ben", "dis_ben", "rent", "fam_allow",
              "house_allow", "cap_inv", "tax_adj")

### estimate a unit model
emdi_model <- emdiplus::ebp(fixed = as.formula(paste("eqIncome ~ ", paste(variables, 
                                                                         collapse= "+"))),
                           pop_data = eusilcA_pop2, 
                           pop_domains = "district", 
                           smp_data = eusilcA_smp2, 
                           smp_domains = "district",
                           na.rm = TRUE,
                           weights = "weight",
                           pop_weights = "popweights",
                           MSE = TRUE,
                           threshold = 11000,
                           B = 2,
                           L = 2)
                           
### full data of highest population below threshold by rank (descending order)
ebp_report_byrank(ebp_object = emdi_model,
                 smp_data = eusilcA_smp2,
                 welfare = "eqIncome",
                 calibvar = "state",
                 domainvar = "district",
                 threshold = 11000,
                 smp_weights = "weight")           
                 
                                   
}
