% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepost_estimation_helpers.R
\name{ebp_reportcoef_table}
\alias{ebp_reportcoef_table}
\title{Produce coefficient table for reporting}
\usage{
ebp_reportcoef_table(ebp_object, decimals = 3)
}
\arguments{
\item{ebp_object}{the EBP object produced from by EMDI from unit model estimation
the object is of class "ebp emdi"}

\item{decimals}{the number of decimals to report on coefficient estimates}
}
\description{
This function takes the object of class 'ebp' to present the regression
model results having specified the number of decimal places.
}
\examples{
data("eusilcA_pop")
data("eusilcA_smp")

#### set of variables used in model estimation
variables <- c("gender", "eqsize", "cash", "self_empl",
              "unempl_ben", "age_ben", "surv_ben",
              "sick_ben", "dis_ben", "rent", "fam_allow",
              "house_allow", "cap_inv", "tax_adj")

### estimate a unit model
emdi_model <- emdiplus::ebp(fixed = as.formula(paste("eqIncome ~ ", paste(variables, 
                                                                         collapse= "+"))),
                           pop_data = eusilcA_pop2, 
                           pop_domains = "district", 
                           smp_data = eusilcA_smp2, 
                           smp_domains = "district",
                           na.rm = TRUE,
                           weights = "weight",
                           pop_weights = "popweights",
                           MSE = TRUE,
                           threshold = 11000,
                           B = 2,
                           L = 2)

ebp_reportcoef_table(emdi_model, 4)

}
