% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepost_estimation_helpers.R
\name{ebp_reportdescriptives}
\alias{ebp_reportdescriptives}
\title{Create Descriptive Statistics for Small Area Estimation Report}
\usage{
ebp_reportdescriptives(
  ebp_object,
  smp_weights,
  pop_weights,
  repvar,
  welfare,
  smp_data,
  pop_data,
  threshold,
  pop_domains,
  smp_domains
)
}
\arguments{
\item{ebp_object}{the EBP object produced from by EMDI from unit model estimation
the object is of class "ebp emdi"}

\item{smp_weights}{the sample weight variable in the household dataset
(i.e. the training data), include column of 1s (DO NOT LEAVE UNSPECIFIED)}

\item{pop_weights}{the population weight variable in the census (or synthetic)
census dataset (i.e. the test data), include column of 1s (NO NOT LEAVE
UNSPECIFIED)}

\item{repvar}{the variable level at which Coefficient of Variation should be
computed}

\item{welfare}{the welfare aggregate variable or outcome variable of interest}

\item{smp_data}{the survey/training data}

\item{pop_data}{the population/census/training data}

\item{threshold}{the poverty line or threshold specified}

\item{pop_domains}{the target area variable within `pop_data`}

\item{smp_domains}{the target area variable within `smp_data`}
}
\description{
This function estimates the coefficient of variation at level specified, basic
statistics such number of units, regions and target areas as well as the
threshold on which SAE is applied and the outcome indicator of interest
(i.e. poverty line and poverty rate). These indicators are all expressed for
the census and survey
}
\examples{
data("eusilcA_pop")
data("eusilcA_smp")
#### set of variables used in model estimation
variables <- c("gender", "eqsize", "cash", "self_empl",
              "unempl_ben", "age_ben", "surv_ben",
              "sick_ben", "dis_ben", "rent", "fam_allow",
              "house_allow", "cap_inv", "tax_adj")

### estimate a unit model
emdi_model <- emdiplus::ebp(fixed = as.formula(paste("eqIncome ~ ", paste(variables, 
                                                                         collapse= "+"))),
                           pop_data = eusilcA_pop2, 
                           pop_domains = "district", 
                           smp_data = eusilcA_smp2, 
                           smp_domains = "district",
                           na.rm = TRUE,
                           weights = "weight",
                           pop_weights = "popweights",
                           MSE = TRUE,
                           threshold = 11000,
                           B = 2,
                           L = 2)

### model estimation
ebp_reportdescriptives(ebp_object = emdi_model,
                      smp_weights = "weight",
                      pop_weights = "popweights",
                      repvar = "state",
                      welfare = "eqIncome",
                      smp_data = eusilcA_smp2, 
                      pop_data = eusilcA_pop2, 
                      threshold = 11000, 
                      pop_domains = "district", 
                      smp_domains = "district")
                      


}
