% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepost_estimation_helpers.R
\name{ebp_test_means}
\alias{ebp_test_means}
\title{Perform test for difference between survey and census means}
\usage{
ebp_test_means(smp_data, pop_data, varlist, smp_weights, pop_weights)
}
\arguments{
\item{smp_data}{the survey/training data}

\item{pop_data}{the population/census/training data}

\item{varlist}{character vector, the set of variables of interest 
(ensure all factor variables are converted into integers to avoid error messages)}

\item{smp_weights}{the sample weight variable in the household dataset
(i.e. the training data), include column of 1s (DO NOT LEAVE UNSPECIFIED)}

\item{pop_weights}{the population weight variable in the census (or synthetic)
census dataset (i.e. the test data), include column of 1s (NO NOT LEAVE
UNSPECIFIED)}
}
\description{
This function computes weighted means of the same set of variables within the
census and the survey. A test for difference of the means are performed for
each variable with two-tailed p-values returned.
}
\examples{
data("eusilcA_pop")
data("eusilcA_smp")

#### set of variables used in model estimation
variables <- c("gender", "eqsize", "cash", "self_empl",
              "unempl_ben", "age_ben", "surv_ben",
              "sick_ben", "dis_ben", "rent", "fam_allow",
              "house_allow", "cap_inv", "tax_adj")

### estimate a unit model
emdi_model <- emdiplus::ebp(fixed = as.formula(paste("eqIncome ~ ", paste(variables, 
                                                                         collapse= "+"))),
                           pop_data = eusilcA_pop2, 
                           pop_domains = "district", 
                           smp_data = eusilcA_smp2, 
                           smp_domains = "district",
                           na.rm = TRUE,
                           weights = "weight",
                           pop_weights = "popweights",
                           MSE = TRUE,
                           threshold = 11000,
                           B = 2,
                           L = 2)

### model estimation
ebp_test_means(varlist = variables,
              smp_weights = "weight",
              pop_weights = "popweights",
              smp_data = eusilcA_smp2, 
              pop_data = eusilcA_pop2)


}
